ci: move sparsity test to own file
